requirements.txt
Django==1.8.2
Pillow==2.8.1
PyYAML==3.11
Unidecode==0.04.17
argparse==1.2.1
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.1
django-cors-headers==1.1.0
django-filter==0.10.0
django-imagekit==3.2.6
django-rest-swagger==0.3.1
django-storages==1.1.8
django-suit==0.2.13
django-taggit==0.14.0
django-toolbelt==0.0.1
djangorestframework==3.1.2
ecdsa==0.13
gunicorn==19.3.0
oauthlib==0.7.2
paramiko==1.15.2
pilkit==1.1.12
psycopg2==2.6
pycrypto==2.6.1
pystache==0.5.4
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sqlparse==0.1.15
static==1.1.1
static3==0.6.1
vobject==0.6.6
wsgiref==0.1.2
django-pagination
python-social-auth
humanize
sorl-thumbnail
django-tinymce==1.5.2